passing Eip2333 tests

+++ src/bls/types/ec/JacobianPoint.cs
@@ -115,7 +115,7 @@ public class JacobianPoint
 
     public bool IsValid() => IsOnCurve() && Multiply(Ec.N).Equals(X is Fq ? InfinityG1() : InfinityG2());
 
-    public int GetFingerprint() => Hmac.Hash256(ToBytes()).Take(4).ToArray().BytesToInt(Endian.Big);
+    public long GetFingerprint() => Hmac.Hash256(ToBytes()).Take(4).ToArray().BytesToInt(Endian.Big);
 
     public AffinePoint ToAffine()
     {
